Restruturando projeto

+++ src/main/java/com/generator/core/model/SerieGrafico.java
@@ -8,9 +8,9 @@ public class SerieGrafico implements Serializable {
 	private static final long serialVersionUID = 5399910651884018303L;
 
 	public String nome;
-	public List<Long> valores;
+	public List<Double> valores;
 	
-	public SerieGrafico(String nome, List<Long> valores) {
+	public SerieGrafico(String nome, List<Double> valores) {
 		this.nome = nome;
 		this.valores = valores;
 	}
